+++ internal/integrations/gemini/schema.go
@@ -3,7 +3,10 @@ package gemini
 import (
 	"context"
 	"fmt"
+	"strings"
 
+	"github.com/vlatan/video-store/internal/drivers/rdb"
+	"github.com/vlatan/video-store/internal/models"
 	"google.golang.org/genai"
 )
 
@@ -111,4 +114,27 @@ func (s *Service) responseSchema(ctx context.Context) *genai.Schema {
 		},
 		Required: []string{"summary", "category"},
 	}
+}
+
+// catString creates a string of categories separated by comma
+func (s *Service) catString(ctx context.Context) string {
+
+	// Get the categories from cache or DB
+	categories, _ := rdb.GetCachedData(
+		ctx,
+		s.rdb,
+		"categories",
+		s.config.CacheTimeout,
+		func() (models.Categories, error) {
+			return s.catsRepo.GetCategories(ctx)
+		},
+	)
+
+	// Extract the category names
+	catNames := make([]string, len(categories))
+	for i, cat := range categories {
+		catNames[i] = cat.Name
+	}
+
+	return strings.Join(catNames, ", ")
 }
